+++ benches/oj_join.rs
@@ -6,8 +6,15 @@ use brunch::{
 	Bench,
 	benches,
 };
-use oxford_join::OxfordJoin;
-use std::time::Duration;
+use oxford_join::{
+	Conjunction,
+	OxfordJoin,
+};
+use std::collections::{
+	BTreeMap,
+	BTreeSet,
+	HashSet,
+};
 
 
 
@@ -15,28 +22,45 @@ const ONE: [&str; 1] = ["Apples"];
 const TWO: [&str; 2] = ["Apples", "Bananas"];
 const THREE: [&str; 3] = ["Apples", "Bananas", "Oranges"];
 const FIVE: [&str; 5] = ["Apples", "Bananas", "Oranges", "Pears", "Jackfruit"];
+const THIRTYTWO: [&str; 32] = [
+	"0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C", "D", "E", "F",
+	"G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V",
+];
 const SLICE: &[&str] = &["Apples", "Bananas", "Oranges", "Pears", "Jackfruit"];
 
 
 
-benches!(
-	Bench::new("oxford_and", "([T; 1])")
-		.timed(Duration::from_secs(1))
-		.with(|| ONE.oxford_and()),
+fn main() {
+	let map = FIVE.into_iter().enumerate().collect::<BTreeMap<usize, &str>>();
+	let set = BTreeSet::from(FIVE);
+	let set2 = HashSet::from(FIVE);
 
-	Bench::new("oxford_and", "([T; 2])")
-		.timed(Duration::from_secs(1))
-		.with(|| TWO.oxford_and()),
+	benches!(
+		inline:
 
-	Bench::new("oxford_and", "([T; 3])")
-		.timed(Duration::from_secs(1))
-		.with(|| THREE.oxford_and()),
+		Bench::new("<[T; 1]>::oxford_and()").run(|| ONE.oxford_and()),
+		Bench::new("<[T; 2]>::oxford_and()").run(|| TWO.oxford_and()),
+		Bench::new("<[T; 3]>::oxford_and()").run(|| THREE.oxford_and()),
+		Bench::new("<[T; 5]>::oxford_and()").run(|| FIVE.oxford_and()),
+		Bench::new("<[T; 32]>::oxford_and()").run(|| THIRTYTWO.oxford_and()),
 
-	Bench::new("oxford_and", "([T; 5])")
-		.timed(Duration::from_secs(1))
-		.with(|| FIVE.oxford_and()),
+		Bench::spacer(),
 
-	Bench::new("oxford_and", "(&[T])")
-		.timed(Duration::from_secs(1))
-		.with(|| SLICE.oxford_and())
-);
+		Bench::new("<&[T]>::oxford_and()").run(|| SLICE.oxford_and()),
+
+		Bench::spacer(),
+
+		Bench::new("BTreeMap::<_, T>::oxford_and()").run(|| map.oxford_and()),
+		Bench::new("BTreeSet::<T>::oxford_and()").run(|| set.oxford_and()),
+
+		Bench::spacer(),
+
+		// HashSet doesn't implement OxfordJoin directly.
+		Bench::new("Conjunction::And.oxford_join(&HashSet<T>)")
+			.run(|| Conjunction::And.oxford_join(&set2)),
+
+		Bench::spacer(),
+
+		Bench::new(r#"<[T; 32]>::join(", ")"#).run(|| THIRTYTWO.join(", ")),
+	);
+}