+++ src/lib.rs
@@ -2,6 +2,7 @@
 # Oxford Join
 
 [![Documentation](https://docs.rs/oxford_join/badge.svg)](https://docs.rs/oxford_join/)
+[![Changelog](https://img.shields.io/crates/v/oxford_join.svg?label=Changelog&color=9cf)](https://github.com/Blobfolio/oxford_join/blob/master/CHANGELOG.md)
 [![crates.io](https://img.shields.io/crates/v/oxford_join.svg)](https://crates.io/crates/oxford_join)
 [![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](https://github.com/Blobfolio/oxford_join)
 
@@ -76,6 +77,10 @@ extern crate alloc;
 
 use alloc::{
 	borrow::Cow,
+	collections::{
+		BTreeSet,
+		BTreeMap,
+	},
 	string::String,
 	vec::Vec,
 };
@@ -87,6 +92,11 @@ use core::{
 
 
 
+/// # Comma + Space.
+const COMMASPACE: &[u8] = b", ";
+
+
+
 #[derive(Debug, Copy, Clone, Default, Eq, Hash, PartialEq)]
 /// # Conjunction.
 ///
@@ -94,8 +104,8 @@ use core::{
 ///
 /// If you're doing something weird and the preset entries aren't currint it
 /// for you, you can use [`Conjunction::Other`], which wraps an `&str`. This
-/// value should just be the word/symbol; any outer whitespace is provided
-/// during join.
+/// value should just be the word/symbol; surrounding whitespace and
+/// punctuation are added during the join as needed.
 ///
 /// ## Examples.
 ///
@@ -183,20 +193,6 @@ impl Conjunction<'_> {
 	/// # Length.
 	///
 	/// Return the string length of the conjunction.
-	///
-	/// ## Examples
-	///
-	/// ```
-	/// use oxford_join::Conjunction;
-	///
-	/// for i in [
-	///     Conjunction::Ampersand, Conjunction::And,
-	///     Conjunction::AndOr, Conjunction::Nor,
-	///     Conjunction::Or, Conjunction::Plus
-	/// ] {
-	///     assert_eq!(i.as_str().len(), i.len());
-	/// }
-	/// ```
 	pub const fn len(&self) -> usize {
 		match self {
 			Self::And | Self::Nor => 3,
@@ -212,20 +208,89 @@ impl Conjunction<'_> {
 	///
 	/// An empty conjunction makes no sense, but because `Conjunction::Other`
 	/// wraps arbitrary values, it is worth checking.
-	///
-	/// ## Examples
-	///
-	/// ```
-	/// use oxford_join::Conjunction;
-	///
-	/// assert_eq!(Conjunction::And.is_empty(), false);
-	/// assert_eq!(Conjunction::Other("foo").is_empty(), false);
-	/// assert_eq!(Conjunction::Other("").is_empty(), true);
-	/// ```
 	pub const fn is_empty(&self) -> bool {
 		match self {
 			Self::Other(s) => s.is_empty(),
 			_ => false,
+		}
+	}
+}
+
+impl Conjunction<'_> {
+	/// # Oxford Join (Generic).
+	///
+	/// This convenience method allows you to Oxford-join _any_ iterable data
+	/// source that yields `AsRef<str>`.
+	///
+	/// If your data type implements [`OxfordJoin`], you should use its
+	/// [`OxfordJoin::oxford_join`] implementation instead as that will be
+	/// faster — they're _specialized_ — but you'll get the same result either
+	/// way.
+	///
+	/// ## Examples
+	///
+	/// ```
+	/// use oxford_join::{Conjunction, OxfordJoin};
+	/// const LIST: [&str; 3] = ["Apples", "Bananas", "Carrots"];
+	///
+	/// // Arrays, for example, implement OxfordJoin, so you should leverage
+	/// // the trait method instead.
+	/// assert_eq!(LIST.oxford_join(Conjunction::And), "Apples, Bananas, and Carrots");
+	///
+	/// // But if you use this method anyway, you'll get the same answer:
+	/// assert_eq!(Conjunction::And.oxford_join(LIST), "Apples, Bananas, and Carrots");
+	///
+	/// // A more appropriate use case for this method would be something like
+	/// // the following:
+	/// assert_eq!(
+	///     Conjunction::And.oxford_join("hello".chars().map(|c| c.to_string())),
+	///     "h, e, l, l, and o"
+	/// );
+	/// ```
+	pub fn oxford_join<I, T>(&self, iter: I) -> String
+	where T: AsRef<str>, I: IntoIterator<Item=T> {
+		iter.into_iter().collect::<Vec<_>>().oxford_join(*self).into_owned()
+	}
+}
+
+impl Conjunction<'_> {
+	/// # Append for Three+.
+	///
+	/// This writes the conjunction with a leading comma-space and trailing
+	/// space to the buffer, e.g. `", and "`.
+	fn append_to(&self, v: &mut Vec<u8>) {
+		match self {
+			Self::Ampersand => { v.extend_from_slice(b", & "); },
+			Self::And => { v.extend_from_slice(b", and "); },
+			Self::AndOr => { v.extend_from_slice(b", and/or "); },
+			Self::Nor => { v.extend_from_slice(b", nor "); },
+			Self::Or => { v.extend_from_slice(b", or "); },
+			Self::Other(s) => {
+				v.extend_from_slice(COMMASPACE);
+				v.extend_from_slice(s.as_bytes());
+				v.push(b' ');
+			},
+			Self::Plus => { v.extend_from_slice(b", + "); },
+		}
+	}
+
+	/// # Append for Two.
+	///
+	/// This writes the conjunction with a leading and trailing space to the
+	/// buffer, e.g. `" and "`.
+	fn append_two(&self, v: &mut Vec<u8>) {
+		match self {
+			Self::Ampersand => { v.extend_from_slice(b" & "); },
+			Self::And => { v.extend_from_slice(b" and "); },
+			Self::AndOr => { v.extend_from_slice(b" and/or "); },
+			Self::Nor => { v.extend_from_slice(b" nor "); },
+			Self::Or => { v.extend_from_slice(b" or "); },
+			Self::Other(s) => {
+				v.push(b' ');
+				v.extend_from_slice(s.as_bytes());
+				v.push(b' ');
+			},
+			Self::Plus => { v.extend_from_slice(b" + "); },
 		}
 	}
 }
@@ -254,10 +319,10 @@ impl Conjunction<'_> {
 /// assert_eq!(set.oxford_join(Conjunction::And), "Apples");
 ///
 /// let set = ["Apples", "Oranges"];
-/// assert_eq!(set.oxford_join(Conjunction::And), "Apples and Oranges");
+/// assert_eq!(set.oxford_join(Conjunction::Or), "Apples or Oranges");
 ///
 /// let set = ["Apples", "Oranges", "Bananas"];
-/// assert_eq!(set.oxford_join(Conjunction::And), "Apples, Oranges, and Bananas");
+/// assert_eq!(set.oxford_join(Conjunction::AndOr), "Apples, Oranges, and/or Bananas");
 /// ```
 pub trait OxfordJoin {
 	/// # Oxford Join.
@@ -327,89 +392,133 @@ pub trait OxfordJoin {
 }
 
 impl<T> OxfordJoin for [T] where T: AsRef<str> {
+	#[allow(unsafe_code)]
 	/// # Oxford Join.
 	fn oxford_join(&self, glue: Conjunction) -> Cow<str> {
 		match self.len() {
 			0 => Cow::Borrowed(""),
 			1 => Cow::Borrowed(self[0].as_ref()),
-			2 => Cow::Owned(join_two(self[0].as_ref(), self[1].as_ref(), glue)),
-			n => {
-				let glue = glue.as_str();
-				let len = glue.len() + 1 + ((n - 1) << 1) + slice_len(self);
+			2 => {
+				let a = self[0].as_ref().as_bytes();
+				let b = self[1].as_ref().as_bytes();
 
-				let (last, rest) = self.split_last().unwrap();
-				let mut base: String = String::with_capacity(len);
-				for s in rest {
-					base.push_str(s.as_ref());
-					base.push_str(", ");
+				let len = a.len() + b.len() + 2 + glue.len();
+				let mut v = Vec::with_capacity(len);
+				v.extend_from_slice(a);  // First.
+				glue.append_two(&mut v); // Conjunction.
+				v.extend_from_slice(b);  // Last.
+
+				// Safety: strings in, strings out.
+				let out = unsafe { String::from_utf8_unchecked(v) };
+				Cow::Owned(out)
+			},
+			n => {
+				let last = n - 1;
+				let len = glue.len() + 1 + (last << 1) + self.iter().map(|x| x.as_ref().len()).sum::<usize>();
+				let mut v = Vec::with_capacity(len);
+
+				// Write the first.
+				v.extend_from_slice(self[0].as_ref().as_bytes());
+
+				// Write the middles.
+				for s in &self[1..last] {
+					v.extend_from_slice(COMMASPACE);
+					v.extend_from_slice(s.as_ref().as_bytes());
 				}
 
-				base.push_str(glue);
-				base.push(' ');
+				// Write the conjunction and last.
+				glue.append_to(&mut v);
+				v.extend_from_slice(self[last].as_ref().as_bytes());
 
-				base.push_str(last.as_ref());
-
-				Cow::Owned(base)
+				// Safety: strings in, strings out.
+				let out = unsafe { String::from_utf8_unchecked(v) };
+				Cow::Owned(out)
 			},
 		}
 	}
+}
+
+impl<T> OxfordJoin for [T; 0] where T: AsRef<str> {
+	#[inline]
+	/// # Oxford Join.
+	///
+	/// This is a special case; the result is always empty.
+	fn oxford_join(&self, _glue: Conjunction) -> Cow<str> { Cow::Borrowed("") }
 }
 
 impl<T> OxfordJoin for [T; 1] where T: AsRef<str> {
 	#[inline]
 	/// # Oxford Join.
 	///
-	/// This is a special case; the only array entry will be returned as-is.
+	/// This is a special case; the sole entry will be returned as-is.
 	fn oxford_join(&self, _glue: Conjunction) -> Cow<str> {
 		Cow::Borrowed(self[0].as_ref())
 	}
 }
 
 impl<T> OxfordJoin for [T; 2] where T: AsRef<str> {
+	#[allow(unsafe_code)]
 	#[inline]
 	/// # Oxford Join.
 	///
 	/// This is a special case; it will always read "first <CONJUNCTION> last".
 	fn oxford_join(&self, glue: Conjunction) -> Cow<str> {
-		Cow::Owned(join_two(self[0].as_ref(), self[1].as_ref(), glue))
+		let a = self[0].as_ref().as_bytes();
+		let b = self[1].as_ref().as_bytes();
+
+		let len = a.len() + b.len() + 2 + glue.len();
+		let mut v = Vec::with_capacity(len);
+		v.extend_from_slice(a);  // First.
+		glue.append_two(&mut v); // Conjunction.
+		v.extend_from_slice(b);  // Last.
+
+		// Safety: strings in, strings out.
+		let out = unsafe { String::from_utf8_unchecked(v) };
+		Cow::Owned(out)
 	}
 }
 
-/// # Join Arrays.
+/// # Join Arrays (3+).
 macro_rules! join_arrays {
 	($($num:literal $pad:literal $last:literal),+ $(,)?) => ($(
 		impl<T> OxfordJoin for [T; $num] where T: AsRef<str> {
+			#[allow(unsafe_code)]
 			/// # Oxford Join.
 			fn oxford_join(&self, glue: Conjunction) -> Cow<str> {
-				let glue = glue.as_str();
-				let len = glue.len() + $pad + slice_len(self.as_slice());
+				let len = glue.len() + $pad + self.iter().map(|x| x.as_ref().len()).sum::<usize>();
+				let [first, mid @ .., last] = self;
+				let mut v = Vec::with_capacity(len);
 
-				let mut base: String = String::with_capacity(len);
-				for s in self.iter().take($last) {
-					base.push_str(s.as_ref());
-					base.push_str(", ");
+				// Write the first.
+				v.extend_from_slice(first.as_ref().as_bytes());
+
+				// Write the middles.
+				for s in mid {
+					v.extend_from_slice(COMMASPACE);
+					v.extend_from_slice(s.as_ref().as_bytes());
 				}
 
-				base.push_str(glue);
-				base.push(' ');
+				// Write the conjunction and last.
+				glue.append_to(&mut v);
+				v.extend_from_slice(last.as_ref().as_bytes());
 
-				base.push_str(self[$last].as_ref());
-
-				Cow::Owned(base)
+				// Safety: strings in, strings out.
+				let out = unsafe { String::from_utf8_unchecked(v) };
+				Cow::Owned(out)
 			}
 		}
 	)+);
 }
 
 join_arrays!(
-	3 5 2,
-	4 7 3,
-	5 9 4,
-	6 11 5,
-	7 13 6,
-	8 15 7,
-	9 17 8,
-	10 19 9,
+	 3  5  2,
+	 4  7  3,
+	 5  9  4,
+	 6 11  5,
+	 7 13  6,
+	 8 15  7,
+	 9 17  8,
+	10 19  9,
 	11 21 10,
 	12 23 11,
 	13 25 12,
@@ -434,71 +543,149 @@ join_arrays!(
 	32 63 31,
 );
 
+macro_rules! join_btrees {
+	($iter:ident) => (
+		#[allow(unsafe_code)]
+		/// # Oxford Join.
+		fn oxford_join(&self, glue: Conjunction) -> Cow<str> {
+			match self.len() {
+				0 => Cow::Borrowed(""),
+				1 => Cow::Borrowed(self.$iter().next().unwrap().as_ref()),
+				2 => {
+					let mut iter = self.$iter();
+					let a = iter.next().unwrap().as_ref().as_bytes();
+					let b = iter.next().unwrap().as_ref().as_bytes();
 
+					let len = a.len() + b.len() + 2 + glue.len();
+					let mut v = Vec::with_capacity(len);
+					v.extend_from_slice(a);  // First.
+					glue.append_two(&mut v); // Conjunction.
+					v.extend_from_slice(b);  // Last.
 
-/// # Slice Length.
-fn slice_len<T>(src: &[T]) -> usize
-where T: AsRef<str> {
-	src.iter().map(|x| x.as_ref().len()).sum()
+					// Safety: strings in, strings out.
+					let out = unsafe { String::from_utf8_unchecked(v) };
+					Cow::Owned(out)
+				},
+				n => {
+					let last = n - 1;
+					let len = glue.len() + 1 + (last << 1) + self.$iter().map(|x| x.as_ref().len()).sum::<usize>();
+
+					let mut v = Vec::with_capacity(len);
+					let mut iter = self.$iter();
+
+					// Write the first.
+					v.extend_from_slice(iter.next().unwrap().as_ref().as_bytes());
+
+					// Write the middles. (Last is count minus one, but since
+					// we already wrote an entry, we need to subtract one
+					// again.)
+					for s in iter.by_ref().take(last - 1) {
+						v.extend_from_slice(COMMASPACE);
+						v.extend_from_slice(s.as_ref().as_bytes());
+					}
+
+					// Write the conjunction and last.
+					glue.append_to(&mut v);
+					v.extend_from_slice(iter.next().unwrap().as_ref().as_bytes());
+
+					// Safety: strings in, strings out.
+					let out = unsafe { String::from_utf8_unchecked(v) };
+					Cow::Owned(out)
+				},
+			}
+		}
+	);
 }
 
-#[allow(unsafe_code)]
-/// # Join Two.
-fn join_two(a: &str, b: &str, glue: Conjunction) -> String {
-	let a = a.as_bytes();
-	let b = b.as_bytes();
-	let glue = glue.as_str().as_bytes();
+impl<K, T> OxfordJoin for BTreeMap<K, T> where T: AsRef<str> { join_btrees!(values); }
 
-	let mut v: Vec<u8> = Vec::with_capacity(a.len() + b.len() + 2 + glue.len());
-	v.extend_from_slice(a);
-	v.push(b' ');
-	v.extend_from_slice(glue);
-	v.push(b' ');
-	v.extend_from_slice(b);
-
-	// Safety: all inputs were valid UTF-8, so the output is too.
-	unsafe { String::from_utf8_unchecked(v) }
-}
+impl<T> OxfordJoin for BTreeSet<T> where T: AsRef<str> { join_btrees!(iter); }
 
 
 
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use alloc::boxed::Box;
 	use brunch as _;
+
+	const CTEST: [Conjunction; 7] = [
+		Conjunction::Ampersand,
+		Conjunction::And,
+		Conjunction::AndOr,
+		Conjunction::Nor,
+		Conjunction::Or,
+		Conjunction::Other("Boo"),
+		Conjunction::Plus,
+	];
 
 	#[test]
 	fn t_fruit() {
-		// Make sure arrays, slices, vecs, and boxes are treated equally.
+		// Make sure arrays, slices, vecs, boxes, etc., all work out the same
+		// way.
 		macro_rules! compare {
 			($($arr:ident, $expected:literal),+ $(,)?) => ($(
-				assert_eq!($arr.oxford_and(), $expected);
-				assert_eq!($arr.as_slice().oxford_and(), $expected);
-				let (v, b) = two_ways($arr.as_slice());
-				assert_eq!(v.oxford_and(), $expected);
-				assert_eq!(b.oxford_and(), $expected);
+				assert_eq!($arr.oxford_and(), $expected, "Array.");
+				assert_eq!($arr.as_slice().oxford_and(), $expected, "Slice.");
+
+				let v = $arr.to_vec();
+				assert_eq!(v.oxford_and(), $expected, "Vec.");
+				assert_eq!(v.into_boxed_slice().oxford_and(), $expected, "Box.");
+
+				let v: BTreeMap<usize, &str> = $arr.into_iter().enumerate().collect();
+				assert_eq!(v.oxford_and(), $expected, "BTreeMap.");
+
+				let v = BTreeSet::from($arr);
+				assert_eq!(v.oxford_and(), $expected, "BTreeSet.");
 			)+);
 		}
 
-		let arr1: [&str; 1] = ["Apples"];
-		let arr2: [&str; 2] = ["Apples", "Bananas"];
-		let arr3: [&str; 3] = ["Apples", "Bananas", "Carrots"];
-		let arr4: [&str; 4] = ["Apples", "Bananas", "Carrots", "Dates"];
-		let arr5: [&str; 5] = ["Apples", "Bananas", "Carrots", "Dates", "Eggplant"];
+		const ARR0: [&str; 0] = [];
+		const ARR1: [&str; 1] = ["Apples"];
+		const ARR2: [&str; 2] = ["Apples", "Bananas"];
+		const ARR3: [&str; 3] = ["Apples", "Bananas", "Carrots"];
+		const ARR4: [&str; 4] = ["Apples", "Bananas", "Carrots", "Dates"];
+		const ARR5: [&str; 5] = ["Apples", "Bananas", "Carrots", "Dates", "Eggplant"];
+		const ARR32: [&str; 32] = [
+			"0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C", "D", "E", "F",
+			"G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V",
+		];
 
 		compare!(
-			arr1, "Apples",
-			arr2, "Apples and Bananas",
-			arr3, "Apples, Bananas, and Carrots",
-			arr4, "Apples, Bananas, Carrots, and Dates",
-			arr5, "Apples, Bananas, Carrots, Dates, and Eggplant",
+			ARR0, "",
+			ARR1, "Apples",
+			ARR2, "Apples and Bananas",
+			ARR3, "Apples, Bananas, and Carrots",
+			ARR4, "Apples, Bananas, Carrots, and Dates",
+			ARR5, "Apples, Bananas, Carrots, Dates, and Eggplant",
+			ARR32, "0, 1, 2, 3, 4, 5, 6, 7, 8, 9, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, and V",
 		);
 	}
 
-	fn two_ways<'a>(src: &'a [&'a str]) -> (Vec<&'a str>, Box<[&'a str]>) {
-		let v = src.to_vec();
-		let b = Box::from(src);
-		(v, b)
+	#[test]
+	fn conjunction_len() {
+		for c in CTEST {
+			assert_eq!(c.len(), c.as_str().len());
+			assert!(! c.is_empty());
+		}
+
+		assert!(Conjunction::Other("").is_empty());
+	}
+
+	#[test]
+	#[allow(unsafe_code)]
+	fn conjunction_append() {
+		for c in CTEST {
+			// Two.
+			let s = [" ", c.as_str(), " "].concat();
+			let mut v = Vec::new();
+			c.append_two(&mut v);
+			assert_eq!(v, s.as_bytes());
+
+			// Three+.
+			let s = [", ", c.as_str(), " "].concat();
+			v.truncate(0);
+			c.append_to(&mut v);
+			assert_eq!(v, s.as_bytes());
+		}
 	}
 }