+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
 
+
+## [0.2.5](https://github.com/Blobfolio/oxford_join/releases/tag/v0.2.5) - 2022-09-11
+
+### New
+
+* impl `OxfordJoin` for `BTreeSet`
+* impl `OxfordJoin` for `BTreeMap` (values)
+* `Conjunction::oxford_join`
+
+### Changed
+
+* Minor performance improvements
+
+
+
 ## [0.2.4](https://github.com/Blobfolio/oxford_join/releases/tag/v0.2.4) - 2022-08-11
 
 ### Changed